finished anime table
