+++ app/api/aws_helpers.py
@@ -5,8 +5,9 @@ import uuid
 
 BUCKET_NAME=os.environ.get("S3_BUCKET")
 S3_LOCATION = f"https://{BUCKET_NAME}.s3.amazonaws.com/"
+
 # add the extensions
-ALLOWED_EXTENSIONS = {"png", "jpg", "jpeg", "gif"}
+ALLOWED_EXTENSIONS = {"mkv", "mp4", "jpeg", "gif"}
 
 
 s3 = boto3.client(
