--- tlcode/app/config.py
@@ -1,5 +1,0 @@
-import os
-
-
-class Config:
-    SECRRET_KEY = os.environ.get('SECRET_KEY')
