--- tlcode/app/routes/anime.py
@@ -1,7 +1,0 @@
-from flask import Blueprint, jsonify, redirect, request
-from flask_login import login_required
-from app.models.models import db
-
-anime = Blueprint('anime',__name__)
-
-@anime.route('/')
