--- tlcode/app/routes/user.py
@@ -1,9 +1,0 @@
-from flask import Blueprint, jsonify, redirect, request
-from flask_login import login_required
-from app.models.models import db
-
-user = Blueprint('user', __name__)
-
-@user.route('/signin')
-def swag():
-    pass
